Create separate contract and service types with the runtime instance (#1778)
* Refactor SDK Wasm view tests

Test the view APIs only once, because it is shared between the contract
and the service.

* Don't use `system_api` in `ContractStateStorage`

Remove the related system APIs, because they aren't really needed.

* Create separate types for contracts

Keep the state as a separate type from the contract type, and require
methods in `Contract` to create the contract type from the state and to
return the state.

* Add a `runtime` field to the contract types

Allow accessing the contract runtime from the contracts.

* Remove `runtime` parameter from `Contract` methods

The type can be obtained with a default constructor and should be stored
in the contract type instead.

* Restrict how `ContractRuntime` is constructed

Ensure it is only created inside `linera-sdk`.

* Don't use `system_api` in `ServiceStateStorage`

Remove the related system APIs, because they aren't really needed.

* Create separate types for services

Keep the state as a separate type from the service type, and require a
method in `Service` to create the service hnadler type from the state.

* Remove `async_trait` from `Service` trait

The service runs in a single thread, so there's no need for its methods
to be `Send` or `Sync`.

* Add a `runtime` parameter to `Service::new`

Allow accessing the service runtime from the service implementations.

* Remove `runtime` parameter from `handle_query`

The type can be obtained with a default constructor and should be stored
in the service type instead.

* Restrict how `ServiceRuntime` is constructed

Ensure it is only created inside `linera-sdk`.

* Update Wasm fuel test fixtures

Rebuild the Wasm binaries to use the updated SDK and update the expected
fuel consumption values.

* Refactor `check_account_authentication` methods

Simplify their signatures to have less parameters.

* Allow new project templates to include warnings

Tweak CI tests to ignore the warnings. They can be useful for new users
as a guide on what to do next.