Replace execution outcome types with new runtime APIs (#1822)
* Fix feature gate for `KeyPair::generate` method

Ensure it is not included in Wasm application builds.

* Rename `resources` field to `grant`

Use a name different from the type name that also provides more context
to what it is used for.

* Move `OutgoingMessage` type to `linera-base`

Allow the type to be accessed from `linera-execution` as well as from
user applications.

* Don't derive `Clone` for `ApplicationStatus`

Prepare to add a `RawExecutionOutcome` field to it.

* Implement conversion from outgoing message types

Allow converting from the public facing type to the internal type.

* Add a `send_message` API to the host runtime

Keep track of the current executing application's `RawExecutionOutcome`
and allow adding messages to it. When execution finishes, append the
messages to the final outcome.

* Add a `send_message` system API

Allow sending messages from Wasm applications using the WIT interface.

* Add conversion from `ChannelName` to `Destination`

Make it straightforward to use a `ChannelName` as an argument for a
parameter that's typed as `impl Into<Destination>`.

* Add a `ContractRuntime::send_message` method

Use a `MessageSender` type to actually send the message, allowing some
options to be configured using a builder pattern.

* Update examples to send messages with new API

Use the new `send_message` runtime API to send messages, instead of
returning them inside the `ExecutionOutcome`.

* Add a `subscribe` API to the host runtime

Add the subscription request to the executing application's
`RawExecutionOutcome` and when execution finishes append the requests
to the final outcome.

* Add a `subscribe` system API

Allow subscribing to application channels from Wasm applications using
the WIT interface.

* Add a `ContractRuntime::subscribe` method

Allow using the runtime type to request to subscribe to application
channels.

* Use `ContractRuntime::subscribe` in Social example

Update the example application to use the new runtime API.

* Add an `unsubscribe` API to the host runtime

Add the unsubscription request to the executing application's
`RawExecutionOutcome` and when execution finishes append the requests
to the final outcome.

* Add an `unsubscribe` system API

Allow unsubscribing to application channels from Wasm applications using
the WIT interface.

* Add a `ContractRuntime::subscribe` method

Allow using the runtime type to request to subscribe to application
channels.

* Use `ContractRuntime::unsubscribe` in Social app.

Update the example application to use the new runtime API.

* Simplify application return types

Remove the `ExecutionOutcome` and `ApplicationCallOutcome` values from
the returned result.

* Remove the outcome types from the WIT interface

Update the contract entrypoints to return simpler types.

* Remove outcome types from SDK

They are no longer used.

* Remove context types from SDK

They were no longer needed, being replaced by the `ContractRuntime` and
`ServiceRuntime` types.

* Use `runtime.send_message` in execution tests

Instead of adding it to the `RawExecutionOutcome`, in preparation to
remove the outcome type from the return value.

* Remove outcome types from main `Contract` trait

Simplify the return types in the host.

* Update Wasm fuel test fixtures

Ensure the Wasm binaries use the updated entrypoint APIs.

* Rename `OutgoingMessage` into `SendMessageRequest`

Differentiate it from `linera_chain::data_types::OutgoingMessage`.