Wait for refund messages before final assertions in Matching Engine end-to-end test (#1873)
* Add a helper function to obtain the chain tip hash

Performs a GraphQL query to obtain the certificate hash of the tip block
of a requested chain, at least as known by the node service.

* Add a `wait_for_messages` helper method

Queries the node service until the sender block has been received and
persisted in the node service. The goal is to use this to ensure that a
block sending messages to the chains tracked in the node service has
been received and processed, so that the messages have been properly
received.

* Fix Matching Engine end-to-end test race condition

Update Matching Engine test to wait for messages before asserting the
final balances, to ensure that the node services have received all
tokens when queried.

Otherwise there was a race between the query for the assertion and the
notification sent from a validator for the block in the admin chain that
sends the refund messages.