Build Witty Wasm test modules sooner in CI
Now that `linera-witty` has been added as a dependency of
`linera-execution`, it seems like the tests now enable the `wasmer`
feature of `linera-witty`. This causes the tests to fail because the
necessary Wasm modules weren't built until later in the pipeline.

This commit moves the compilation of those Wasm modules to before the
tests are executed.