Fix incorrect delivery notifications for rejected blocks (#1933)
* Fix incorrect message delivery notification

If an error occurred while handling the cross-chain request to confirm
that the recipient was updated then the block is rejected. The updates
to the chain state are rolled back correctly, but the delivery notifiers
can't be rolled back, so they may have notified the delivery of a
message incorrectly.

The solution is to run the delivery notifier only after all operations
that can fail have succeeded.

* Remove check to see if change state changed

If it hasn't changed, it's still safe to save it. Removing this
optimization makes the code simpler, and easier to read and audit.