Rename `new` and `finalize` to `load` and `store` (#1994)
* Take ownership of `self` in `finalize`

Make it clearer that this is called at the end of the transaction.

* Rename `new` to `load`

Use a name that pairs well with the intended new name for `finalize`:
`store`. It also reflects better what the developer should include in
the method's implementation.

* Rename `Contract::finalize` to `Contract::store`

Make it clearer what the main intention of the method is, so that new
developers have a clearer sense of what must be done at the end of the
transaction.