Pin `async-graphql` version (#2012)
* Update and pin the version of `async-graphql`

The latest version (7.0.5) unfortunately breaks our code, because of an
arguably semver incompatible change. Therefore, the latest version that
works is pinned.

* Also update the `Cargo.toml.template` file

Ensure that the projects created by `linera project new` don't fail
because of the arguable semver incompatibility with `async-graphql`
version 7.0.5.

* Ignore that `async-graphql-derive` isn't used

It was added in order to force pin its version so that it matches the
version of `async-graphql`.