PublishBlob (#2014)
## Motivation

Currently users are able to upload bytecode that can be used to create applications. Other chains have to subscribe to the publisher chain to be able to use the bytecode. We want to generalize this so that:
* Users can publish arbitrary data that can be used as bytecode or read by existing applications.
* One doesn’t have to subscribe to the publisher chain (and wait for it to handle the subscription request) to use the blob.
* Blobs are content-addressed, so that if two users upload identical blobs, they are stored only once.

## Proposal

This PR implements the `PublishBlob` operation

## Test Plan

Unit test

## Release Plan

Since we're adding a new `SystemOperation`, I'm guessing we have to do this:
- Need to update the developer manual.