Listen for signals in the main task (#2086)
There's no need to spawn a separate task for it, because as soon as the
signals are registered, any received signals lead to events being sent
to the streams. And since the channel used to notify that a shutdown
signal was received is never dropped, the future waiting for a signal
can just be polled directly.