Work around stack overflow
The `change_voting_rights::rocks_db` test in `linera-core` was failing
because it overflowed its stack. While trying to investigate why, this
refactor seemed to fix it. This should be good enough for now, and since
this part of the code will be refactored so that it runs inside a
separate task running a `ChainWorkerActor`, the stack pressure should be
reduced then.