Replace request_sent_certificates_in_range and use the download of certificates (#2091)
## Motivation

Right now we have `request_sent_certificates_in_range` in the chain info query. We don't need to go all the way to the server to get these certificates.

## Proposal

With the new endpoints, as discussed on Slack, we could modify this to return just the certificate hashes, and then download the certificates from the proxy directly using the new endpoints.

## Test Plan

CI