Track the certified block that published a blob last (#2156)
## Motivation

For example, if we know a certain blob was published, and we try to use it, but some of the validators don't have it, we need to be able to know the certificate that last used the blob so that we can sync validators to also have that blob. That's one use case for this, but there are more. Part of #2033

## Proposal

Track what `Certificate` last used the blob. As of right now, the only "usage" is publishing a blob, but in the future we'll also use it in the `read_blob` system API call, when we replace `Bytecode` with blobs, etc.

## Test Plan

CI + will use this in following PRs