Retrieve pending messages in `execute_block`
Messages were always retrieved before `execute_block` anyway, but in a
way that left the possibility of a race condition. The race could happen
when there was a pending block, because `execute_block` would propose
the pending block before the requested block. That means that there
would be some repeated incoming messages in the pending block and the
requested block. With this refactor, the list of incoming messages is
prepared after the pending block has been proposed.