Make `SyncRuntime` represent the execution lifetime (#2192)
* Refactor to create `SyncRuntimeHandle`

Rename the `SyncRuntime` type into `SyncRuntimeHandle` and create a new
`SyncRuntime` wrapper type with the goal of representing the execution
lifetime. It can then be able to clear the loaded applications when it
is dropped.

* Move `into_inner` to `SyncRuntime`

Instead of in `SyncRuntimeHandle`, so that it better matches the
lifetime of the loaded applications.

* Clear applications when `SyncRuntime` is dropped

Ensure that there are no memory leaks.

* Remove deprecated manual clearing of applications

No longer needed now that the destructor clears the loaded applications.

* Fix conjugation in function documentation

Follow the convention used everywhere else.

* Refactor to split `create_dummy_application`

Will be used to create fake applications with fake instances, that
aren't contracts or services.

* Refactor to split `create_contract_runtime`

Allow creating generic runtimes, with a custom generic type.

* Test if loaded applications are cleared on drop

Check that it doesn't leak memory.

* Test if `into_inner` avoids leaking memory

It should return `None` if it detects there are still loaded
applications, which would reference the runtime.

* Test if `into_inner` works after cleaning up

It should succeed and not leak memory if loaded applications have been
cleared.