Run service runtime as an actor spawned by the chain worker (#2209)
* Replace `new` method with `From` implementation

Prepare to add `new` methods that are specific to the concrete
implementation.

* Import `QueryContext` in `runtime.rs`

Instead of using the longer `crate::QueryContext` path.

* Refactor to split `ServiceSyncRuntime::run_query`

Prepare to allow reusing an instance of the runtime to handle multiple
queries.

* Prepare to run service runtime as an actor

It needs to run on its own thread, but it needs to handle multiple
queries. Therefore, it should be executed in a separate thread and run
like an actor, handling one request at a time.

* Run the service runtime as an actor

Send the query request from outside the runtime thread.

* Rename `Request` type to `ExecutionRequest`

Make it clearer to which actor the requests are for.

* Document `ChainStateView::chain_id` method

Ensure that there's some documentation available.

* Add `ChainWorkerState::current_query_context`

Prepare to have the `ChainWorkerActor` capable of changing the context
when needed.

* Move service runtime actor thread to chain worker

Prepare to keep the thread alive across multiple queries.