read_blob system API (#2193)
## Motivation

We need a `read_blob` system API so that users are able to use blobs that were already published before. This is part of #2073

## Proposal

This implements the `read_blob` system API, handling the oracle records properly for it as well. For this PR we're not dealing with lagging validators, we'll do that in a following PR. For now if a validator doesn't have the blob, it'll just fail the block proposal.

## Test Plan

Added a test for when we try to read a blob that wasn't published yet, and making sure we fail in that case.