Ensure only data blobs are present in SDK/System API calls (#2368)
## Motivation

We don't want `BlobId` exposed to the SDK. Users will only be aware of `Data` blobs, the other types will be only internally used.

## Proposal

Only expose `Data` blobs to the SDK, and assume when dealing with blobs that it's a `Data` blob. Because of that we don't need to use `BlobId` everywhere, and can just use the hash of the `Data` blob in most places.

## Test Plan

CI + also retested the NFT UI