Expand env variables inside gql in READMEs (#2400)
## Motivation

Harcoding this type of thing is generally not great, as it makes the READMEs break more easily.

## Proposal

Add the hability of expanding env variables inside the graphQL queries. Also store the query result in an env variable, in case we want to actually use that result in the README (which is already the case in a few of them).
Fixes #2256

## Test Plan

CI