Replace Bytecode with Blobs (#2375)
## Motivation

Now that we have blobs, there are several follow ups we can do.

## Proposal

This PR replaces Bytecode with Blobs. This will clean up a lot of code and also make our blocks smaller when publishing Bytecodes. Fixes #2149 #2199

## Test Plan

CI
Some tests had to be altered since we don't have to subscribe to get bytecodes anymore, and also the `ApplicationId` format has changed with this.