Avoid evicting in-use `ChainWorkerActor`s
A race condition was possible where the cache became full and the chain
worker that was evicted was still running. This is usually not problem,
but it is a problem if a new `ChainWorkerActor` is created for the same
chain before the previous one has finished handling its pending
requests.

Eviction is now more careful, and selects the worker to evict based on
the LRU strategy and only considers the workers that aren't active. The
check for if it's active is done based on the endpoint's strong count,
because all queries to the worker keep the sender endpoint alive while
waiting for a response (see: `WorkerState::query_chain_worker`).