Compress and decompress bytecodes in blocking thread (#2471)
* Decompress contract bytecode in a blocking thread

Instead of blocking one of Tokio's asynchronous worker threads.

* Decompress service bytecodes in a blocking thread

Instead of blocking one of Tokio's asynchronous worker threads.

* Compress bytecodes in a blocking thread

Instead of blocking one of Tokio's asynchronous worker threads.

* Remove incorrect TODO

`ruzstd` doesn't support compression, so we can't use it for adding
compression. Compression isn't really needed for the Wasm target,
because we can compress and publish bytecodes from outside the browser.

* Replace `.into()` with `.compress()`

Make it explicit that the operation is not trivial.

* Replace `.try_into()` with `.decompress()`

Make it explicit that the operation is not trivial.