Use a new `StubInstance` to avoid requiring the "test" feature (#2509)
* Move `wit_generation` module into a sub-directory

Prepare to add a new sub-module with a `StubInstance`, so that there's
no need to use the `MockInstance`.

* Create a new `StubInstance` type

A replacement for the `MockInstance` type to use when generating WIT
files. This allows generating the files without having to activate the
"test" feature.

* Use the `StubInstance` in the `wit_generator`

Replace usages of the `MockInstance` type and don't enable
`linera-witty`'s "test" feature.