Make BlobContent into enum (#2465)
## Motivation

Right now we have `BlobContent` always as a blob of bytes, but it can have many forms. This causes us to have to serialize/deserialize into those forms back and forth in the code more than we have to.

## Proposal

Encode the blob type into the `BlobContent`. Now `BlobId` and `BlobContent` will have the type encoded in them. They're supposed to be a 1:1 mapping to each other. This will avoid unecessary serialization/deserialization

## Test Plan

CI