Add unit test for saving wallet with pending blobs (#2679)
## Motivation

CI was failing on PR #2646 with a JSON serialization failure. That was fixed by PR #2668 but we realized none of our tests covered this case.

## Proposal

Create a test that tries to save a wallet with pending blobs in it. Verified that before #2668 this test fails, and now it succeeds.

## Test Plan

CI

## Release Plan

- Nothing to do / These changes follow the usual release cycle.