Push all tags individually (#2780)
## Motivation

There seems to be some evidence online that `docker push -a` doesn't seem to be 100% reliable. We're seeing some images not have all the tags they're supposed to have.

## Proposal

`docker push` each tag individually, so we know for sure everything is being pushed and tagged properly.

## Test Plan

Will push this to `devnet_2024_10_21`, which was the branch that we saw had this issue in the latest image.

## Release Plan

- These changes should be backported to the latest `devnet` branch
- These changes should be backported to the latest `testnet` branch