Remove blobs cache (#2784)
## Motivation

This cache isn't needed, and actually was incorrect in how we were using it: the cache wrongfully technically gave access to unpublished blobs to any chain in the client, as well as masked a bug (we needed to be checking the chain manager's pending blobs in `get_blobs` also).

## Proposal

Remove the cache

## Test Plan

Existing ts in CI should pass and confirm that this indeed wasn't needed.

## Release Plan
I feel like this should probably be backported everywhere?

- These changes should be backported to the latest `devnet` branch
- These changes should be backported to the latest `testnet` branch