Test receipt of old epoch change messages (#2797)
* Move import to start of file

It will also be used by a new test.

* Test receipt of old epoch change messages

Epoch change messages are protected, and can't be rejected. Therefore,
old epoch change messages must be allowed to be received, even if they
are simply ignored.

* Test receipt of old epoch deprecation messages

Ensure that past messages to remove committees can also be accepted by a
chain that is already in a future epoch.