Fix lint CI and split it into multiple jobs (#2785)
## Motivation

On #1737 we added `continue-on-error` on the job level. AFAIU that will succeed even if the entire jobs fails with a cancellation for example, which will happen when we timeout.
This was causing this job to silently succeed while timing out for a while.

## Proposal

Split the different lints into their own jobs, adding parallelism and also giving us individual failures

## Test Plan

CI

## Release Plan

- Nothing to do / These changes follow the usual release cycle.