Don't use glob imports for fieldless enums (#2853)
## Motivation

For fieldless enums, if we use glob imports before a match statement, and then remove one of the enum variants, that removed enum variant silently becomes a catch all in the match statement.
The compiler shows no errors for it.

## Proposal

Stop using glob imports for fieldless enums. Now removing a variant should cause a compilation error for these, as expected.

## Test Plan

CI

## Release Plan

- Nothing to do / These changes follow the usual release cycle.