client_for_proxy_worker -> worker_client (#2893)
## Motivation

A lot of the code around this function changed, and I think we forgot to also update the name

## Proposal

Change it from `client_for_proxy_worker` to `worker_client`. This is already inside `GrpcProxy`, so I think having `proxy` in the name is redundant. The name of the struct that this returns is `ValidatorWorkerClient`, so `worker_client` seemed to make the most sense.
I think (IMHO) this makes it a lot easier to understand what this function is returning.

## Test Plan

CI

## Release Plan

- Nothing to do / These changes follow the usual release cycle.