[summary] Track CI job runtime (#2936)
## Motivation

It would be nice to see CI runtime regressions or improvements in the PRs.

## Proposal

Create Rust CLI tool that checks runtimes, compares against base ref, and comments on the PR with a summary of the changes.
This works, but as you guys can see it's pretty ugly for now, we're just posting the JSON 😅
We're also commenting for every successful workflow run in CI.

The PR is already pretty big so I will do the following follow ups:
- [x] Markdown format this in a pretty way (done in #2946)
- [x] Add the performance summary markdown as a [job summary](https://github.blog/news-insights/product-news/supercharging-github-actions-with-job-summaries/) of the job, so people have access to historical versions of this (done in #2948)
- [x] Find if there's already a performance summary comment, and edit it instead of always posting a new one (done in #2953)
- [ ] Might make sense to only post if the change exceeds some percentage threshold for now. Open to suggestions

Other follow ups are tracked in the issue for this: #2834

## Test Plan

Triggered a modified version of this manually with workflow dispatch, it works, but we can unfortunately only know for sure this works by merging :/ 

## Release Plan

- Nothing to do / These changes follow the usual release cycle.