[summary] Update existing comment (#2953)
## Motivation

Right now the performance summary always "spams" the PR with comments, one for everytime the `Rust` workflow finishes successfully in CI.

## Proposal

Now that we have the summary attached to the performance summary CI job, we have a way of accessing the historical values for these things. So now we can just edit an existing comment instead of always creating a new one.

## Test Plan

Will be tested with workflow dispatch, like in previous PRs. You can see that there's only one comment, and it's edited (instead of just commenting again)

## Release Plan

- Nothing to do / These changes follow the usual release cycle.