Add `State` suffix to template's state type (#2987)
* Add `State` suffix to template's state type

The `Contract` and `Service` traits are no longer implemented for that
type, so the type no longer represents the application itself. Make it
clearer that the type just holds the application's persisted state.

* Add `State` suffix to AMM state type

Use the new convention of adding a `State` suffix to the suffix types.

* Add `State` suffix to Counter state type

Use the new convention of adding a `State` suffix to the suffix types.

* Add `State` suffix to Crowd-Funding state type

Use the new convention of adding a `State` suffix to the suffix types.

* Add `State` suffix to Etherum-Tracker state type

Use the new convention of adding a `State` suffix to the suffix types.

* Add `State` suffix to Fungible Token state type

Use the new convention of adding a `State` suffix to the suffix types.

* Add `State` suffix to GenNFT state type

Use the new convention of adding a `State` suffix to the suffix types.

* Remove `state` module from LLM example

It is not used anywhere.

* Add `State` suffix to Matching Engine state type

Use the new convention of adding a `State` suffix to the suffix types.

* Add `State` suffix to NonFungible Token state type

Use the new convention of adding a `State` suffix to the suffix types.

* Add `State` suffix to Social state type

Use the new convention of adding a `State` suffix to the suffix types.