Subscriptions, mutations, and the `counter` example (linera-io/linera-web#16)
* Restructure repository

The client library we expect users to load in Web frontends is built
alongside the browser extension, so the communication protocol between
that library and the extension is now encapsulated: the only public
API we need to keep stable is the TypeScript API of the library, and
developers can pull in the library using their JavaScript package
manager of choice.

The client worker is now its own package independent of the extension:
this should make it easier to build demos like the where the client is
embedded directly into the frontend and the user gets a throwaway
wallet to play with.

* Subscriptions and mutations

* Use release build of Linera

* Add counter example

* CI with Clippy

---------

Co-authored-by: Janito Vaqueiro Ferreira Filho <janito.vff@gmail.com>