Update compose-server-init.sh
The set -e option has been enabled, which terminates the script execution upon the first error.

A log_message function has been added to write logs to a file with a timestamp. This helps track the process and identify the causes of errors.


Parameters such as the storage address and the configuration file path can be easily modified by setting them as environment variables.

All messages are now sent to both the log and the console for easier tracking of the process.

Signed-off-by: gluttony <107846669+Kinaqu@users.noreply.github.com>