Track certificates signed by each validator (#3038)
## Motivation

We want to track validators that are potentially lagging behind (not signing as many blocks).

## Proposal

Add a Prometheus counter that counts for each validator, how many of the confirmed certificates we have processed they signed.
If we really want to keep an eye on this, we could even add an alert later on if the validator with the most signatures minus the validator with the least signatures is greater than some threshold.

## Test Plan

Did a local run and checked the metrics:

![Screenshot 2024-12-17 at 23.23.23.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/HlciHFAoHZW62zn13apJ/b9386742-d093-4cc9-808d-25683fe687d5.png)

## Release Plan

- These changes should be backported to the latest `testnet` branch