Simplify blobs: only type and bytes. (#3062)
## Motivation

In anticipation of application and chain description blobs I had pushed
for making `BlobContent` an enum. However, we need the bytes for the
size limit check, and in many cases we won't need the deserialized
content at all: E.g. when we are a validator and only send the blob to a
client.

## Proposal

Make blobs a simple pair of a `BlobType` and bytes (`[u8]`).

## Test Plan

This does not add any new features yet. CI should catch regressions.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- Closes https://github.com/linera-io/linera-protocol/issues/3050.
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)