Implement `WitType`, `WitLoad` and `WitStore` for pointer types (#3063)
## Motivation

<!--
Briefly describe the goal(s) of this PR.
-->
It's not uncommon to want to store some values in the heap, but still be
able to copy them to and from guest Wasm modules. For that to work, heap
smart pointer types should implement the `WitType`, `WitLoad` and
`WitStore` traits.

## Proposal

<!--
Summarize the proposed changes and how they address the goal(s) stated
above.
-->
Create a new macro with the code to delegate implement the traits by
delegating the items to the wrapped type, and use the macro to implement
the traits for `Box`, `Rc`, and `Arc`.

## Test Plan

<!--
Explain how you made sure that the changes are correct and that they
perform as intended.

Please describe testing protocols (CI, manual tests, benchmarks, etc) in
a way that others
can reproduce the results.
-->
Unit tests were modified to wrap all tested types in `Arc`, `Rc` and
`Box` containers and check that it doesn't affect the loading and
storing of the types.
Unit tests were added to check derived `WitType`, `WitLoad` and
`WitStore` implementations for types that have `Arc`, `Rc` and `Box`
fields.

## Release Plan

<!--
If this PR targets the `main` branch, **keep the applicable lines** to
indicate if you
recommend the changes to be picked in release branches, SDKs, and
hotfixes.

This generally concerns only bug fixes.

Note that altering the public protocol (e.g. transaction format, WASM
syscalls) or storage
formats requires a new deployment.
-->
- These changes introduce a backwards compatible feature, so they follow
the usual release cycle.

## Links

<!--
Optional section for related PRs, related issues, and other references.

If needed, please create issues to track future improvements and link
them here.
-->
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)

---------

Signed-off-by: Janito Vaqueiro Ferreira Filho <janito.vff@gmail.com>
Co-authored-by: Andreas Fackler <afck@users.noreply.github.com>